Extract the loading of reviews to iife for reuse
